[hotfix] move fft to operator.common.dataproc
