+++ flink-ml-parent/flink-ml-lib/src/main/java/org/apache/flink/ml/operator/common/dataproc/FFT.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.flink.ml.common.utils;
+package org.apache.flink.ml.operator.common.dataproc;
 
 import org.apache.commons.math3.complex.Complex;
