+++ flink-ml-parent/flink-ml-lib/src/test/java/org/apache/flink/ml/operator/common/dataproc/FFTTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.flink.ml.common.utils;
+package org.apache.flink.ml.operator.common.dataproc;
 
 import org.apache.commons.math3.complex.Complex;
 import org.junit.Test;
